add countdown timer to TimeHelper.cs

+++ src/hurry/Helpers/ResultsHelper.cs
@@ -11,7 +11,7 @@ public static class ResultsHelper
             // to normalize results, we are considering 1 "word" to equal 5 characters, so we determine
             // words entered by dividing total characters by 5
 
-            var wordCount = test.UserInput.Length / 5;
+            var wordCount = test.UserInput.Length / 5L;
 
             // we first calculate words per second
 
@@ -19,7 +19,7 @@ public static class ResultsHelper
 
             // then we multiply words per second by 60 to get words per minute (wpm)
 
-            var results = wordsPerSecond * 60;
+            var results = (int)wordsPerSecond * 60;
             return results;
         }
         catch (Exception)
